+++ frontend/src/App.css
@@ -1,45 +1,5 @@
-/*.App {
-  text-align: center;
-}
 
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}*/
-
-
-
-@import url("https://fonts.googleapis.com/css2?family=Sansita+Swashed:wght@600&display=swap");
+/* @import url("https://fonts.googleapis.com/css2?family=Sansita+Swashed:wght@600&display=swap"); */
 body {
   margin: 0;
   padding: 0;
@@ -48,69 +8,6 @@ body {
   justify-content: center;
   align-items: center;
   height: 100vh;
-  background: linear-gradient(45deg, greenyellow, dodgerblue);
+  /* background: linear-gradient(45deg, greenyellow, dodgerblue); */
   font-family: "Sansita Swashed", cursive;
 }
-.App {
-  position: relative;
-  padding: 50px 50px;
-  background: rgba(255, 255, 255);
-  border-radius: 10px;
-}
-.App h1 {
-  font-size: 2em;
-  border-left: 5px solid dodgerblue;
-  padding: 10px;
-  color: #000;
-  letter-spacing: 5px;
-  margin-bottom: 60px;
-  font-weight: bold;
-  padding-left: 10px;
-}
-.App .form-control {
-  position: relative;
-  /* width: 300px; */
-  height: 50px;
-  margin-bottom: 50px;
-}
-input {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 90%;
-  border: 1px solid rgb(81, 81, 81);
-  outline: none;
-  /* background: none; */
-  padding: 10px;
-  border-radius: 10px;
-  font-size: 1.2em;
-}
-
-.App .form-control label {
-  position: absolute;
-  top: 14px;
-  left: 20px;
-  font-size: 1em;
-  transition: 0.6s;
-  font-family: sans-serif;
-}
-/* .App .form-control input:focus ~ label,
-.App .form-control input:valid ~ label {
-  transform: translateX(-13px) translateY(-35px);
-  font-size: 1em;
-} */
-button {
-  /* width: 30%; */
-  padding: 12px;
-  background: dodgerblue;
-  align-items: center;
-  font-size: 18px;
-  color: #fff;
-  border: #fff;
-  border-radius: 10px;
-}
-button:hover {
-  background: rgb(222, 22, 22);
-  cursor: pointer;
-}
-
