--- frontend/src/components/ParticlesBackground.js
@@ -1,13 +1,0 @@
-import Particles from "react-tsparticles"
-import particlesConfig from "./config/particles-config"
-import { loadFull } from "tsparticles";
-
-const ParticlesBackground = () => {
-    return(
-        <Particles params={particlesConfig}>
-
-        </Particles>
-    )
-}
-
-export default ParticlesBackground
