+++ frontend/src/components/Welcome.js
@@ -1,0 +1,7 @@
+import React from 'react'
+import "./welcome.css"
+export const Welcome = () => {
+  return (
+    <div>Welcome</div>
+  )
+}
