+++ frontend/src/components/inputbox.css
@@ -1,0 +1,125 @@
+
+/* @import url("https://fonts.googleapis.com/css2?family=Sansita+Swashed:wght@600&display=swap"); */
+
+.Box {
+  position: relative;
+  padding: 50px 50px;
+  background: rgba(255, 255, 255);
+  border-radius: 10px;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  width: 50vw;
+  border: 1px solid blue;
+}
+.info{
+    /* border:1px solid red; */
+    width: 40%;
+    background-color: rgba(124, 239, 239, 0.508);
+    padding: 10px;
+    align-items: center;
+    justify-content: center;
+    padding-top:30px;
+    /* border: 1px solid red; */
+    padding-left: 20px;
+}
+
+.form{
+    /* border:1px solid red; */
+    width: 60%;
+    border-left: 3px solid rgb(255, 244, 39);
+    padding-left: 20px;
+    font-family: sans-serif;
+    /* border: 1px solid red; */
+}
+h1 {
+  font-size: 2em;
+  /* border-left: 5px solid dodgerblue; */
+  padding: 10px;
+  color: #000;
+  letter-spacing: 5px;
+  margin-bottom: 60px;
+  font-weight: bold;
+  padding-left: 10px;
+}
+.form-control {
+  position: relative;
+  /* width: 300px; */
+  height: 50px;
+  margin-bottom: 50px;
+  margin-left: 10px;
+}
+input {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 90%;
+  border: 1px solid rgb(81, 81, 81);
+  outline: none;
+  /* background: none; */
+  padding: 10px;
+  border-radius: 10px;
+  font-size: 1em;
+}
+
+.form-control label {
+  position: absolute;
+  top: 14px;
+  left: 20px;
+  font-size: 1em;
+  transition: 0.6s;
+  font-family: sans-serif;
+}
+/* .Box .form-control input:focus ~ label,
+.Box .form-control input:valid ~ label {
+  transform: translateX(-13px) translateY(-35px);
+  font-size: 1em;
+} */
+button {
+  /* width: 30%; */
+  padding: 12px;
+  background: dodgerblue;
+  align-items: center;
+  font-size: 18px;
+  color: #fff;
+  border: #fff;
+  border-radius: 10px;
+}
+button:hover {
+  background: rgb(222, 22, 22);
+  cursor: pointer;
+}
+
+@media screen and (max-width: 644px){
+    .Box{
+        flex-direction: column;
+        /* width: ; */
+    }
+   .form{
+        width: 100%;
+        padding: 0px;
+        margin: 0;
+        border-top: 2px solid yellow;
+        border-left: 0;
+        margin-top: 5px;
+    }
+    .info{
+        width: 100%;
+        padding: 0;
+        margin: 0;
+        padding-left: 10px;
+    }
+    h1{
+        margin-bottom: 10px;
+    }
+    .form-control{
+        /* padding: 0; */
+        /* border: 1px solid red; */
+        margin: 2px;
+    }
+    button{
+        font-size: 15px;
+        padding: 10px;
+    }
+    
+}
